+++ Library/Homebrew/extend/git_repo_path.rb
